+++ src/ndapplib/nd_httpListener.cpp
@@ -74,7 +74,7 @@ static int _session_data_handler(nd_handle sessionHandler, void *data, size_t le
 	if (!pSession) {
 		return -1;
 	}
-	return pSession->onDataRecv((char*)data, len,(NDHttpListener*) NDGetListener(listen_h));
+	return pSession->onDataRecv((char*)data, (int)len,(NDHttpListener*) NDGetListener(listen_h));
 }
